CLDR-19640 CLDRModify support copyNew with regex target
- also, split CLDRConfigFileFilter out because CLDRModify is a large unwieldy file
- add -V verbose mode
